+++ components/WalletSelection/data.ts
@@ -1,0 +1,36 @@
+const WALLETS = [
+  {
+    name: 'Metamask',
+    image: '/static/images/wallets/metamask.png',
+  },
+  {
+    name: 'Rabby Wallet',
+    image: '/static/images/wallets/rabby.png',
+  },
+  {
+    name: 'Coinbase Wallet',
+    image: '/static/images/wallets/coinbase.png',
+  },
+  {
+    name: 'Rainbow',
+    image: '/static/images/wallets/rainbow.png',
+  },
+  {
+    name: 'WalletConnect',
+    image: '/static/images/wallets/wallet-connect.png',
+  },
+  {
+    name: 'Trust Wallet',
+    image: '/static/images/wallets/trust-wallet.png',
+  },
+  {
+    name: 'Safe Wallet',
+    image: '/static/images/wallets/safe-wallet.png',
+  },
+  {
+    name: 'Ledger Live',
+    image: '/static/images/wallets/ledger-live.png',
+  },
+]
+
+export default WALLETS